♻️ Run bin for pipeline tests (#34)
Installs the packed `css-typed` package from CI Build into a sandbox for
the pipeline tests and switches to using `npx css-typed`. This uses the
actual `bin` entry point instead of emulating it with `node dist/main.js`.

In addition to matching the real world use case better, this ensures that
we do not break the bin configuration again. (It broke in the 0.5.0 release.)